refactor: merge ui/messenger and communication/ui
